add severity string to db search output
